backend/requirements.txt
Flask
Flask-Cors
Pillow
numpy
opencv-python
python-dotenv
scipy
scikit-learn